refactor:DailryPage 로 부터 페이지 이동, 페이지 모달 기능 분리
* PageNavigator
   - 페이지 이동, 페이지 모달을 띄우는 기능 담당
